Melee enemy added, lists of enemy added
Melee enemy has a proper model, I have changed the model because there was a light on it and there was a camera on the weapon. So lights and camera have been removed. Enemy combat can't be polished yet because combat is not working, same for enemy autism

Enemies now spawn with the help of a list, can add however many enemies as you want